+++ yeat/workflow/__init__.py
@@ -21,20 +21,12 @@ def run_workflow(args):
     config = vars(args)
     config["data"] = get_config_data(config["config"])
     config["bandage"] = check_bandage()
-    print(args)
-    print(snakefile)
-    print(config)
-    assert 0
     if args.grid == "slurm":
-        success = snakemake_grid_slurm(
-            snakefile, snakemake_args, config, grid_nodes, log_path, dry
-        )
-    elif args.grid:
-        success = snakemake_grid_default(
-            snakefile, snakemake_args, config, grid_nodes, grid_args, log_path, dry
-        )
+        success = snakemake_grid_slurm(args, snakefile, config)
+    elif args.grid == True:
+        success = snakemake_grid_default(args, snakefile, config)
     else:
-        success = snakemake_local(snakefile, snakemake_args, config, dry)
+        success = snakemake_local(args, snakefile, config)
     if not success:
         raise RuntimeError("Snakemake Failed")  # pragma: no cover
 
@@ -79,77 +71,52 @@ def check_bandage():
 def setup_grid_args(args):
     if args.grid_args is not None:
         return args.grid_args
-    grid_args = " -V "
-    if args.threads > 1:
-        grid_args = f" -V -pe threads {args.threads} "
-    return grid_args
+    grid_args = "sbatch " if args.grid == "slurm" else " -V "
+    thread_arg = f"-c {args.threads} " if args.grid == "slurm" else f"-pe threads {args.threads} "
+    return grid_args + thread_arg
 
 
-def snakemake_local(s_file, s_args, config, dry):
-    # success = snakemake(
-    #     snakefile,
-    #     config=config,
-    #     cores=args.threads,
-    #     dryrun=args.dry_run,
-    #     printshellcmds=True,
-    #     workdir=args.outdir,
-    #     use_conda=True,
-    #     local_cores=args.threads,
-    #     nodes=args.grid_limit,
-    #     drmaa=setup_grid_args(args),
-    #     drmaa_log_dir=str((Path(args.outdir) / "gridlogs").resolve()),
-    # )
+def snakemake_grid_slurm(args, snakefile, config):
     success = snakemake(
-        s_file, **s_args, config=config, cores=config["threads"], dryrun=dry, printshellcmds=True
+        snakefile,
+        config=config,
+        dryrun=args.dry_run,
+        printshellcmds=True,
+        workdir=args.outdir,
+        use_conda=True,
+        local_cores=args.threads,
+        nodes=args.grid_limit,
+        drmaa=setup_grid_args(args),
+        drmaa_log_dir=str((Path(args.outdir) / "gridlogs").resolve()),
+        cluster=args.grid_args,
     )
     return success
 
 
-def snakemake_grid_slurm(s_file, s_args, config, nodes, log_path, dry):
-    # success = snakemake(
-    #     snakefile,
-    #     config=config,
-    #     cores=args.threads,
-    #     dryrun=args.dry_run,
-    #     printshellcmds=True,
-    #     workdir=args.outdir,
-    #     use_conda=True,
-    #     local_cores=args.threads,
-    #     nodes=args.grid_limit,
-    #     drmaa=setup_grid_args(args),
-    #     drmaa_log_dir=str((Path(args.outdir) / "gridlogs").resolve()),
-    # )
+def snakemake_grid_default(args, snakefile, config):
     success = snakemake(
-        s_file,
-        **s_args,
+        snakefile,
         config=config,
-        local_cores=config["threads"],
-        nodes=nodes,
-        drmaa_log_dir=log_path,
-        slurm=True,
-        dryrun=dry,
+        dryrun=args.dry_run,
+        printshellcmds=True,
+        workdir=args.outdir,
+        use_conda=True,
+        local_cores=args.threads,
+        nodes=args.grid_limit,
+        drmaa=setup_grid_args(args),
+        drmaa_log_dir=str((Path(args.outdir) / "gridlogs").resolve()),
     )
     return success
 
 
-def snakemake_grid_default(s_file, s_args, config, nodes, grid_args, log_path, dry):
-    # success = snakemake(
-    #     snakefile,
-    #     config=config,
-    #     cores=args.threads,
-    #     dryrun=args.dry_run,
-    #     printshellcmds=True,
-    #     workdir=args.outdir,
-    #     use_conda=True,
-    # )
+def snakemake_local(args, snakefile, config):
     success = snakemake(
-        s_file,
-        **s_args,
+        snakefile,
         config=config,
-        local_cores=config["threads"],
-        nodes=nodes,
-        drmaa=grid_args,
-        drmaa_log_dir=log_path,
-        dryrun=dry,
+        cores=args.threads,
+        dryrun=args.dry_run,
+        printshellcmds=True,
+        workdir=args.outdir,
+        use_conda=True,
     )
     return success