+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Customize downsample, depth of coverage, and genome size per sample (#77)
 
 ### Changed
-- Updated `--grid` and `--grid-args` arguments and Snakemake to be compatible with SLURM (!78)
+- Updated `--grid` and `--grid-args` arguments and Snakemake to be compatible with SLURM (#79)
 
 
 ## [0.5] 2024-01-08